Huge adds & fixes for 3d models!
MODELS & DEFS:
:new: Fixed ALL windows for 3d vehicles
:new: Fixed transparency to SlopeMe1 & 2 actors (3d slopeable windows),
so now we can place on maps and see the glass from both sides!
:new: New Flakveirling model from CoD2, with a destroyed variant too
:exclamation:
:new: Converted .md2 models into .md3 ones where possible

DECORATE:
:new: Turned the FlakVierling3d actor into a switchable one: now if we
turn it off, we'll see the destroyed variant of the model (with
darksmoke though :heart: )

MAPS - TESTMAP:
:new: Added both FlakVierling3d variants into the map

MAPS - C1M4:
:new: Fixed FlakVierling actors inside the map
:new: Fixed Script 11 accordingly to the new FlakVierling actor (instead
of remove the actor, now deactivate it - also uncommented the Junk spawn
too)

I presume I need to add a nice explosion to the FlakVierling3d so we
don't have to take care though for it while building an ACS (pretty sure
that I'll go for a TanKExplosion)

More has yet to come. :smile: