Updates to stealth code
- Fixed duplicate numbering of global variable that governed heartbeat
sound activation.
- Tweaked amount that crouch influences visibility so that you can't
just run around crouched and be invisible to enemies.
- Added a slight visibility modifier based on skill, so that higher
skill levels can't crouch-run everywhere without breaking the visibility
threshold.
- Added AlertLight actor that increases internal visibility variable
when the player is in range set via argument 1 (default 64).  Range is
constructed to be similar in size to a dynamic light with the same
range.
- Added automatic spawning of AlertLight actors to some static lights
(those used in C2M1).