Tweaks to spotlight actor
- Spotlight will now follow the player (up to a certain angle/radius
from its normal path) if the player is caught in the beam (this was
missing a LOF_NOJUMP flag, but was otherwise already in place - the
actors in C2M1 were also placed oddly, which caused problems.  See
below).
- Tweaked C2M1: The angle of the Spotlight actor should face the same
direction that the light is expected to come from.  Previous iteration
of this map had the actors facing into walls, with the pitch rotated
around to 325+ degrees, but this caused the  line-of-sight calculations
in the Spotlight actor to fail.
- Tweaked SpotlightGlow to force the player to be visible immediately
upon contact with the beam