Minor tweaks to stealth system
- Commented out the settings in the C2M1 map script and changed the
defaults in the library to be the ones that were previously in place in
C2M1.  Map script settings can still override the library, but I don't
think they are needed here.
- Minor script tweaks to make the effect of stealth crouching more
balanced and to update some aspects of standing movement at higher skill
levels.  You still almost *have* to be crouched to play certain areas on
skills 3 and 4, but I don't think that's unreasonable.