SFX: New rain special effect
Now made out of a model, it is possible to pitch it. It also has some
new arguments to choose from as well as two different types for little
or heavy rainfall. The pitched one can be seen in action in C1TRAIN,
C1M2 (coast only ) TESTMAP. Please check the rain maps if everything's
still working as intended