C3M2: Started work on the concentration camp as I imagine it
The old C3M2 is saved as "C3M2_old.wad" for now as I want to take some
references from it. Other than that I think the new version will
resemble the real desolate situation in the kz's of the Third Reich much
closer - at least that's the plan.