C1M5: Polishing
Fixed all missing textures in the Hotel de la Mort.
Fixed HOM/missing textures inside the elevator.
Re-made the underwater slopes.
Re-textured and re-aligned the textures in the front office of the apartments on the side of the bank building.
Removed the allied soldiers from the inside of the bank.
Fixed the HOM/missing textures near the underwater slopes.
Fixed underwater open crate sector heights.

:heavy_exclamation_mark: :heavy_exclamation_mark: Guys, if you want to have BoA released by August 1st, you really should get to work on finishing it! :heavy_exclamation_mark: :heavy_exclamation_mark: