Set compression level to 9 instead of the default 5
With changing the compression level to 9, this will further allow a
tighter squeeze within the overall size of the archive file.  However,
the draw-back will be the amount of processing time it requires to
decompress the data in real-time.  Previously, the value was 5 - as its
the documented as the default value in the 7Zip documentation.