Use the exclusion lists in build.sh.
Rename _zip.bat to build.bat for consistency.
Use the exclusion lists in build.sh.
Change the way 7zExcludeList.txt is used in build.bat.
Change the flag for build.sh to use LZMA or Deflate.
Make the filename/path cases in 7zExcludeList.txt and 7zExcludeListDir.txt match the actual cases of the directories and files. Unix is case sensitive.