Use absolute paths
Use absolute paths to specify all directories and files; this is
necessary to be sure that we are invoking the 'correct' and using the
correct files in-which to compact.  Otherwise, if the user WD's differs
with the program and we try to compact or invoke files or software, we
could accidentally cause Unforeseen Consequences!  (or if 'Tools\7za'
exists in another directory out side of the project, we could compile up
something other than what we want)