Use the 'START' extCMD when invoking 7Zip && Added = %x to make it simple as to what Parameter is what
With using the extCMD START, this can allow some flexibility or atleast
supports the ability to customize how the new process is generally
invoked by the user.  As of right now, it is possible to redefine the
priority level - if coded differently.  But right now, it is hard-coded
at 'Normal'.  Any other value greater than 'Normal' will allow the OS to
provide more execution time towards the 7Zip program as opposed to tasks
with the 'Normal' flag.  However, higher the time given the more
likelyhood that the system in general will lag until the task is
finished.

To make the code and documentation easier to read, I have placed '= %x'
to indicate what parameter is what in the documentation.  This will make
it simpler to follow when reading the code.