Updated ENU: C1 Mission Briefings and Miscellaneous
I've completed editing all of the C1 mission briefings as well as the
introductions for Douglas and Miller. This also includes the secret
mission text and the end of Chapter intermission text (that may have
formatting issues). Although I was able to test some of the formatting
after getting BoA to work again, it wasn't showing my updated text
changes after a while, so there may be issues yet. James has offered to
proof it for me, but if anyone else wants to help, I will edit the ENU.