Updated ENU: System text and NPC dialogue
-Minor revisions to a handful of system texts (quit messages, pick-up
messages, etc.)

-Minor-to-moderate revisions to Laz, Ty, and Juliette's dialogue. In
particular, I spent WAY too much time trying to perfectly balance
Juliette's accent in the dialogue. Finally, I decided to take a minimal
approach and just replace the voiced th- sound (as in "the" and "then"
but not "thin" or "think") with the /z/ sound. I put all the H's back
in, too. Consonant H's are very un-French, but I think her "voice" now,
with just the /z/ thing and the bonjours and mon amis, will still
"sound" French to the player. It should hopefully just be a little
smoother to read at the same time!

One note on Ty. I changed "earnings" to "salary" because that's what I
thought it meant, then I realized, in context, the text could've been
talking about BJ's achievements and accomplishments, not his salary. But
I think the salary thing works well -- it sort of shows the player how
important a character BJ is to the Allied cause, and it fits in the
dialogue because Ty apologizes for his manners afterward :)