Fix inaccessible secret area on C2M0_B.
The said inaccessible secret area is an offshoot of the basement crate room.
Changed the texture on the door leading to the secret area.
Fixed the ceiling textures in the crate room.
Added a switch on the back of one of the monitors upstairs to open the secret area.
Made the door passable.