Revised strings
Corrected directory formatting that is presented to the user.
Display an error code if 7za returned an error during the invoktion
test.
The error messages are mostly generic but it could help end users that
are having issues, this should be easier to track down the issue as
opposed to throwing the 'missing files' detected.  Or in other words, be
sides presenting the user with Windows 8's infamous 'Something Happened'
error screen [ http://rfc1337.net/somethinghappened.png ], try to be a
little more verbose as to what went wrong.