Updated ENU: Lots of work on Ch. 1's briefing texts
Some highlights:

Originally, Blake addressed BJ as either Blazkowicz, BJ, Will, or
William. Since they're close friends, Blake now just calls him Will or
BJ, which are less formal.

Changed Saint-Jean (c1m2 prison) to Saint-Jeanne, so that Blake's joke
in BRIEFINGC1M2_6 works (Jean is a male name in France, so he and BJ
must have had a very wild time indeed! =p). On the first pass I changed
it to Jeanne in his joke, but the joke still didn't quite work because
the pronounciation would be different, so now it's all just
Saint-Jeanne. Because the prison is apparently fictional (or at least,
very non-specific) based on a Google search for it, I thought it would
make more sense to rename the prison rather than delete Blake's joke. If
this is a problem, someone just let me know and I'll change the prison
back to Saint-Jean and replace the joke with something more generic.

In BRIEFINGC1M3_3, changed 56mm to 88mm (the flak guns) for dense
technical reasons--see under "Barrel length" at:
https://en.wikipedia.org/wiki/Caliber_(artillery)
Basically, the 56 is the number of the flak gun's "calibers", plural,
which is apparently a rather abstract unit of measurement used for large
artillery. But the caliber of these guns, in the common sense of the
word, is 88mm.

For c1m5, one of the mission's central objectives is to defend a main
bridge in Paris over the Somme river. It is the Seine that runs through
Paris, so I changed Somme to Seine. I thought the bridge should at least
have a name as well, even if it's only mentioned once during the
briefing text; it might break the player's immersion if the bridge is
only ever referred to as the "main bridge." Thus in the briefing texts I
named it the Pont de Roussel (Roussel is a French name meaning
red-haired, so it's a sort of reference to BJ =p) Again, if that's a
problem, someone let me know and I'll fix it. A note: I only revised the
first three briefing strings for c1m5, given that the map is still being
finalized; I will finish the rest of m5's briefing texts when the
mission's flow is decided on.

Similarly, for the c1m6 briefing, I revised Miller's 5 lines but left
Blake's 2 lines alone -- I'm going to come back to those.

In the c1m0 briefing, I wasn't sure what exactly was meant by "three
sensitive containers" but I assumed these were incubators of sorts for
storing the mutants, so the revisions reflect this.

Tried to fix a lot of the issues with stylistic consistency--the use of
different colors in instances of the same context, Cpt. vs Captain, etc.
For the last one, I've made it so that "Captain" is used when a
character (Douglas or Ryan) is introduced, and Cpt. is used on
successive mentions.