Updated ENU: Lots of work on the intermap texts
Minor revisions for Sgt. Nash at the front and Lt. Dave in the mess
hall. Also made minor revisions on Blake's training dialogue.

Made minor revisions to the new notice about door color. (@Dan - I see
what you were going for with the part about Allied High Command, but I
tried it a few different ways and I think the way it is now works pretty
well. But, as always, let me know if you'd like me to go back to it!)

Softened up the tone of Will and Gutenberg's initial exchange. More
importantly, I also made Gutenberg's explanation of Eisenmann more vague
and less detailed. In the c1m0 briefing, Miller "reveals" the shocking
nature of the Eisenmann experiments, but none of what he reveals is much
more then what Gutenberg says about the operation when you first ask.
Now, Gutenberg mentions only that he isn't sure, but that he believes
the project concerns mutation and cybernetics, and that he needs the
files to learn more. m0's briefing should now have more of an impact on
the player.

The toilet's dialogue still has a way to go, but I'll leave that for the
coders =p I'll come back to it to polish it up. Otherwise, much of the
Ch. 1 intermap text should be just about finished.