Updated ENU: Mostly in-mission text revisions for c1m1 and c1m2
-Minor revisions to in-mission dialogue: Blake in c1m1, Blake & Ryan in
c1m2.

-Minor revisions to Ascher's warning about mines in the intermap.

In c1m1 objectives:
-Changed "Collect yellow key for registration" to "Gain access to
registration building" since I think this is what was originally
meant--"registration" alone is a little unclear. If I'm mistaken,
someone tell me and I'll fix it :) The player will still know that they
need the yellow key because the yellow door appears at the beginning of
the map
-Changed "Dr. Schabbs" to "Schabbs Jr." to make it clear from the start
that the m1 Schabbs is one of the sons.
-Changed "files" to "intel" to prevent any possible confusion between
Eisenmann files and mission files.