Add some more ammo to the start of C1M4, and fix some issues.
Give the player more ammo at the start of C1M4 so that they have 64+32 9mm bullets to start off with.
Fix waterfall scroll direction on the west side of the dam.
Fix texture at the top of the elevator that connects the base interior with the last flakvierling on the outside.