Gray out the mine sweeper power bar if it isn't in use.
Add hidden inventory item which is an indicator of whether or not the mine sweeper is in use. Give said inventory item to the player when they activate their mine sweeper, and take it away when they deactivate it.
Add grayed-out version of the mine sweeper power bar.
Move existing mine sweeper power bar graphics to a new folder.
Change SBARINFO so that it displays the coloured power bar when the player's mine sweeper is on, and the gray power bar when the player's mine sweeper is off.