Attempt to fix objective display errors.
After the player re-loads a quicksave, and another script from boalib runs, that other script would screw with the objectives array for some reason. It's probably a ZDoom bug.

This workaround uses an integer bitfield to tell which objectives are active, and which ones are not. This way, objectives can only be made active, and thus displayed, using the "boaobjectivesset" script.

The objectives display script checks both the bitfield and the objectives array to determine what to show, and what not to show.

This workaround may fix the extraneous objective bug on C1M5.

The workaround works on my end.