Fix Vitality serum at shop, and coffee/water machines.
Make Laz sell you a VitalityToken at the shop instead of a VitalitySerum.
The VitalityToken is a CustomInventory that has a maximum of 10, and it gives you 5 health boosts when you pick it up.
Fix Field kit purchasing at the shop after Operation: Astrostein.
Fix coffee/water machines so they can't be used after they are empty.