Remove VitalityToken, and merge its code with VitalitySerum.
VitalitySerum was just a proxy for UpgradeStaminaBoA, and VitalityToken was just a proxy for VitalitySerum. I think it was pointless making a separate VitalityToken actor just to enforce a limit on the amount you can purchase at the shop, when it's better to replace/rewrite the VitalitySerum, since the VitalitySerum already has dynamic lights, brightmaps, and a DoomEdNum.