Updated ENU: c2m1 in-mission texts & miscellaneous revisions
-Minor revisions to the skill level confirmation messages

-Very minor revisions to a couple of Blake's Ch. 1 texts, for
consistency of style (coloring, etc.)

-Minor revisions to the c2m1 objectives and moderate revisions to
Blake's c2m1 in-mission texts. The fact that Blake is conveniently
waiting for BJ creates a minor plot hole (i.e., why does he happen to be
there?). But the revisions should address that; initially, Blake comes
to the camp to scope the area out and try to come up with a plan to
rescue BJ (DARRENMESSAGE22), then realizes that BJ is breaking out on
his own and waits for him