Fix minesweeper pickup behaviour bug.
I noticed a bug with the minesweeper in this Let's Play video: https://youtu.be/ISuLN49O_jc.
The bug shown in the Let's Play video was that the player had a MineSweeperActive hidden in his inventory, so he ended up getting a coloured minesweeper power bar, indicating that the minesweeper was in use, when it really wasn't.
This made me realize that picking up another minesweeper while having one in your inventory is undefined behaviour.
The minesweeper won't get picked up if the player has a minesweeper in his inventory, and if he hasn't already turned it on.
I've made it so that the minesweeper gets thrown away and deactivated once the player picks up another one, if the player's current minesweeper was used.