Conversation fixes for C1M5 and INTERMAP
Change Juliette's conversation at the end of the level, so that she has something else to say after the player has defeated the Nazi waves.
Also, add a couple of panzerschreck rockets so that the battle with the tank near the end is not so unforgiving if the player misfires.
Remove secret area tag from sectors in C2M1 that can't be registered as secret areas (the snow lump at (X: 6039, Y: -4113)).
Fix C2 INTERMAP conversation with Prof. Gutenberg, so that the player is not prompted for Eisenmann files or to ask about Operation: Eisenmann.
BOALIB: Remove AktenV2 (V2 files) from player's inventory.