C2M3: Fix doors in two-storey house.
The doors in the two-storey house blocked the entire height of the house. This commit fixes that.

TECHNICAL NOTES

The doors' control sectors were previously only as tall as the height of the door. This made it impossible for the player to walk above or underneath the door directly.

The control sectors are now as tall as the house. Combined with the 3D midtex flag on the door polyobject linedefs, this allows the player to walk above and below them.

However, I cannot figure out how to prevent the player from opening a door on the upper floor from the lower floor.