Re-worked solution from previous commit
- Reverted to default color ranges
- Dug out source files for Wolf3D CONFONT and recompiled it at a darker
level so that the default color ranges look better in the console/menus.
- Co-opted an unused glyph on the font for the scrollbar body image
- Tweaked menu to use a modified color "MenuBlack" that defines a
different color range for the console translation so that the menu
cursor shows up red (it was almost invisible after the font
replacement).  This lets the main menu cursor match the conversation
menu cursor (either color can be changed easily).