Player Tracer Changes
- Changed PlayerFollowers to use PlayerTracer shots insteaf of enemy
shots
- ZScriptified the base Tracer classes so that I could us PostBeginPlay
on the PlayerTracer class to set species and flags before the first tic
(because with FastProjectiles, setting in the first tic of spawn is too
late - the first step of movement and damage has already occured)
- Added PostBeginPlay function override to the PlayerTracer that sets
the species to Player if the shooter is a PlayerFollower and
PlayerFollower if the shooter is anything else (a player).
- Added logic so that PlayerFollowers can always shoot through players
and players can shoot through PlayerFollowers on skills 0 and 1 in
single-player games.