Fix PlayerFollower Friendly Fire
- Changed PlayerFollowers to use PlayerTracer shots instead of enemy
shots
- ZScriptified the base Tracer class so taht I could use PostBeginPlay
on the PlayerTracer class to set species and flags before the first tic
-- With FastProjectiles, setting these in the Spawn state was too late;
the first movement step and damage had already occurred
- Added PostBeginPlay override to the PlayerTacer that sets the species
to Player if the shooter is a PlayerFollower, and PlayerFollower if the
shooter is anything else (a player).
- Added logic so that PlayerFollwoers can always shoot through players,
and players can shoot through PlayerFollowers on skills 0 and 1 in
single-player games.
- This also fixes being able to walk through players...  Also added code
to make them walk away if they are too close to you (needs tweaking,
still).