Truck windows follow truck on rail script
- ZScriptified the truck window actor
-- Set actor to tic *after* ACS script execution so that the truck
model's position will have been adjusted before the window tries to move
to the location.
-- Self-moving actors don't have this issue; in this case the script was
moving the parent actor *after* the windows had been drawn, so there was
a 1 tic delay that caused the movement to be off.
- Returned Truck actor definition to original state, with (temporary?)
handling for args[0] being 2 or 3 so levels won't break.
- For even better position matching, the Warp command in the
TruckSequence should have its WARPF_INTERPOLATE flag removed in all of
the maps where this script is used...  Doesn't seem to affect the game
experience, and the occasional slight gap in the windows disappears.