Added visibility indicators to hud views
- Not 100% happy with this, but it's an initial implimentation
- Automatically shows up if a sneakable enemy is in the current map
- Automatically disappears when the last sneakable enemy is dead
- Shows visibility level.  When disguised, shows visibility level in a
subdued color.
- When an actor that can see through NOTARGET is looking at the player,
the visibility level begins to fade back in as red, based on current
suspicion level.
- If the suspicion level is greater than the visibility level, the red
bar shows the suspicion level instead
- Right now, disguises disappear forever if Gestapo see you...  This
should probably be re-worked so that if no enemies are currently looking
for the player, the disguise returns.  It looks like sometimes the ACS
script for incrementing active enemies isn't firing properly, so that
needs work, too.