Added Episode Icon Handling
* All icons are placeholders, obviously - no idea what final images need
to look like*
- Handled with the same code as the skill menu, using some different
handling for the icon drawing/positioning (so description text is
available too, if wanted)
- Horizontal offset adjustments are handled automatically, based on the
width of the largest icon that is found
- Vertical spacing of the menu has to be updated via menudef, depending
on icon height (can't do via ZScript because the Episode list is built
in native code, still)
- Icons (and description text, if wanted) are added following a similar
naming convention as the skills...  e.g., EPISODE1, EPISODE2, EPISODE3,
etc. for the graphics, and EPISODEDESCRIPTION1, EPISODEDESCRIPTION2,
etc. for episode description text (or leave out, and they won't be
drawn, as you can see with this commit)