C2M1: Fixed first portal, added visibility handling to volumetric light cones
- Portal needed to be Static, not Interactive so that enemies are aware
of it and can see through it
- This has the added effect of having the three executioner SSMP40
actors come after you if you actually throw the grenade that you get in
the first room with the scientists/mechanics...  Not sure if we want
that, but it'll add some realism - If you're dumb enough to cause
explosions, someone will come investigate.  Alternative is to script
toggling their AMBUSH flag when you enter/leave the portal area.
- Also added visibility increase handling to the volumetric cone, since
the wall lights didn't add to visibility - It's not a big increase, so
no real gameplay effect other than standardization.