C2M1: Minor tweaks; Also changed how alerted sneakables are counted
- Fixed some crate texture misalignments and a line that I accidentally
had set to be triggerable multiple times.
- Re-worked the alerted enemy counter to just count the number of
SneakableGuardEyesAlerted actors are on the map instead of depending on
manually incrementing/decrementing the counter.  Fixes count issues if a
sneakable actor is active when it is Thing_Remove'd (which could happen
with the three initial guards if you maneuvered them into the portal
area, then woke them up just before you grabbed the cyan key).