Some ZScript tweaks for framerate
- These changes make C2M5 run at almost double the framerate for me
(which, with all effects turned on, is a whopping 11-12 FPS or more now
in most places - up from 5-6 average).
- I still get slowdown on the train and when looking down the central
missile corridor/shaft toward where the 500 radius subtractive light is
- I think simply because of the 8000+ unit distance.  Framerate there
drops to 5-6 for me (which is better than the 3-4 it was before).

Overall though, I can play through most of the level with no issues now.

I think we're running into an actual mapping limitation here...  16,000
thinkers may just too many for the engine to easily handle on most
machines.  I was able to eliminate a bunch by not spawning the sneak
visibility detection part of the dynamic lights when there aren't
sneakable actors in the level on load, so it looks like it's down to
14,000+ now...  Disabling rain/snow/etc. pushes it down to 13,000, and I
can play with few problems.

My earlier attempt to freeze actors that weren't in range didn't work...
Well, it worked, but managed to make things slower despite having fewer
actors active, even with the distance check only running once a second
or so.