C2M5_B: Attempt to fix some of the issues I've noticed
Set tag for all rain spawners and rain sounds within the launch bay to 290 so they get deactivated once the launch bay doors are closed.
Refactor the DECORATE code for all Electricity and ElectricOrb actors so that it is more efficient, and mappers can use arg 0 to specify whether or not a particular electricity is harmless or not.