MauserAmmo drop tweaks
Forced all MauserAmmo drops to give the full 5 rounds.

The Mauser ammo drop sprite is of a full stripper clip (5 rounds).
There's not really a good reason why there would be a partial stripper
clip, even if we added a modified sprite (the metal clip is removed when
you load the rounds into the weapon - normally all 5 at once).