Fixes for sneakable enemies
- Patrolling sneakable enemies could walk up to the player and alert,
but most of the time would then just walk around the player without
opening fire.  Now they will open fire as soon as they alert.
- Eliminated the need to use the ACS portion of patrol resuming - Now,
if a patrolling enemy loses the player, they will immediately go back to
trying to reach the patrol target they were walking toward when they saw
the player.
- Added: non-patrolling sneakable enemies that lose the player now
attempt to return to their original locations.  Doesn't work when
NOTARGET is on (or a disguise is in use with no weapon raised)