Initial commit of working alarms
- Adds 'Alarm' and 'AlarmSpawner' actors for placement on map.
- 'Alarm' actor activation occurs when any sneakable actor within a 512
unit radius around the alarm is alerted.  Deactivation occurs when no
alerted sneakable actors remain within range.
- 'AlarmSpawner' actors will spawn 3 SneakableSSMP40Guard actors at the
spawner's location when the alarm is activated.  The spawner tracks the
guards so that there are never more than 3 alive per spawner, so if two
are killed, two more are spawned the next time the alarm activates, if
none are killed, none are spawned, etc.
- Alarms with no TID will activate all other alarms in the map.
- Alarms and AlarmSpawners can be grouped by giving them a TID in common
(so an alarm with TID 1 will trigger other alarms with TID 1 and any
AlarmSpawners with TID 1, but not anything with TID 2).
- These are only really designed to work around sneakable actors right
now - they'll alert non-sneakables as well, but will deactivate
immediately if there are no sneakables around.