Alarm Updates (Work In Progress)
- Added 'AlarmPanel' actor that should be placed in the map and tagged
with a TID matching that of at least one Alarm actor and/or
AlarmSpawner.
- Changed handling so that the closest sneakable actor to an alarm panel
will walk to the panel and activate it after the actor becomes alerted.
Activation doesn't just happen because sneakables are awake with this
change.
- The alarm will stay active until it is either turned off via script
(Set to 'Deactivate' state) or is manually deactivated at the AlarmPanel
by the player - sneakable actors in range of Alarm actors will continue
to be active until the AlarmPanel is deactivated, then they will return
to Idle as they normally would.
- Non-sneakable actors can still be awakened by the alarm, but will
never go back to being idle.

This is still WIP - I just need to get the code committed.