Generified some components of alarm handling
- Added generic ActorSpawner that is a configurable version of the
AlarmSpawner actor that allows the map author to set the type of actor
spawned (default is Guard), how many (max) spawned actors should be
alive at any given time (default 1), the TID that the spawned actors
should have (default none), and the TID of a PatrolPoint that the
spawned actors should walk to before resuming normal behavior (default
none).
- Also handles *not* spawning when the player can see the spawn point,
or, optionally, when the player is within a certain radius of the
spawner (default 512 units).

All parameters are set via user_* properties in UDMF, except for the
spawned class type, which uses arg0str (this part works, just not in
GZDB currently - I had to edit the textmap manually to test - Posted in
the GZDB-BF thread asking about the problem).

This also helped me discover a couple of minor bugs in the Alarm
handling, which are also fixed now.