Preliminary Interaction Marker Work
- Prep for automating placement and positioning of excalamation point
markers on NPCs.
- New markers become greyed out when the player begins a conversation
with the NPC actor.
- Added inheritance from Base to the Dummy_Talk actors and a few of the
older Statist actors that aren't based off of Nazi actors.

NOTE: There is currently no way to automate the marks showing up when an
actor has an attached conversation ID...  this implementation depends on
the actor having the 'user_conversation' UDMF property set to 1 to
enable showing the marker.

These changes have no effect on the current implementation, so we can
hold off on any map changes until we can automate most of the work...