Minor ZScript Optimizations
- Tweaked the Alarm Panels to use ThinkerIterators even less often
- Tweaked static enemies to only call FindState once per actor

Honestly, I don't know why there would be any significant performance
change as a result of either of the original pieces of code (doesn't
mean it's not happening, I just don't know why it would, and I don't see
it on my end).  The calls happen so rarely that, if anything, any
slowdown would be during the first few seconds after level load, then be
unnoticeable.