Made AMBUSH-flagged Nazis use 'Aim' frames
- Basically, AMBUSH-flagged Nazis act and look just like static Nazis
until they see the player, then they immediately open fire and go to
normal chasing.
- There *might* be some non-standard actors left where this starts them
out in a wierd frame.  I had to add a 'Missile.Aimed' label to the
Grenadiers so they don't randomly show up in a death frame (it was using
the base class's Missile.Aimed frame letter with the grenadier sprite
set)...  But the rest that I tested were fine.
- Also corrected spelling of Grenadier throughout the mod (sorry, Ozy,
it was bothering me :smiley: ).