ZScript Tweaks - Medic, Ryan Playerfollower
- Made Ryan Follower have BUDDHA flag and to heal himself when needed
when he's back next to you and has no target.  Side effect of this is
that he now actually will enter his pain state, so won't just stand
there being shot anymore.
- Tweaked medic heal handling slightly more...  Enemies don't need to be
customized to "wait" for the healing to finish anymore, and multiple
medics should no longer go after the same dead enemy (can still happen
if they 'see' it on the same tic, but not as often as before).