AlarmPanel Tweaks
- Allow alarm panels to properly activate/deactivate any other actor
(e.g., DORMANT actors, etc.)
- Make sound control of alarm more granular, so that it starts as soon
as the alarm panel is activated, and stops as soon as it's deactivated.
- Alarms now skip zero-speed actors when finding an activator (so
Snipers will never activate an alarm)
- C2M1 - added alarm panels to the outdoor sniper/spotlights area, and
tied all machine guns/alarms into them.  Now spotlights only trigger the
alarm if one or more of the three snipers is still alive, and the player
can turn the alarm off from any of the three sniper towers.