Make shake shader work the way I want it to, and add it to C2M4
Now, the shake shader shakes the screen based on how much of the shader control item you have. It assumes the amount you have is in ACS fixed point + 2.

I have made several changes to the "TruckSequence" script on C2M4. The script now repeatedly calculates the difference in position between the railMover's current position and it's previous position, and sets the result as the amount of ShakeShaderControls the player has.