QuickKick Script/NullWeapon tweaks/fixes
- Fixed the quick kick's weapon restore check so that you can still kick
even if your other weapons are all out of ammo and can't be switched to
(added a timeout so that trying to switch back to a wepon with no ammo
doesn't just wait endlessly).
- It was easiest to see this at the start of C2M1, if you fired your
pistol until it ran out of ammo - with an empty weapon, you could kick
once, then would be stuck with no weapon at all.
- Also added "NullWeapon" to the player's start items, ZScriptified it,
and added inventory checks for DisguiseToken handling.  This means that
the player can choose to lower their weapno at any time, with the
NOTARGET piece only kicking in if they have a disguise.  NullWeapon is
set to be a very, very high priority number, so will only be
auto-switched to if there are no other weapons with ammo available - so
normally the knife will take precedence - but it WILL show up when
cycling through weapons, so I re-added it's tag being displayed and
updated that as well.