Implemented Sneakable Actor Weapon Drops
- Sneakable Actors will now *always* drop their weapon, but with a
significantly reduced amount of ammo ( it works out to 1-2 rounds for
pistols and 4 or so for MP40s)
- To do this, I ported some code form my Star Trek mod that allows
setting the dropped weapon's AmmoGive amount by using the drop amount
value (so weapons act similarly to ammo/inventory drops)
- The drop amount is by default used to set the AmmoGive1 amount.  If
the weapon is not set up to give any of AmmoType1, but gives AmmoType 2
(like most of the weapons in Nazis were),  AmmoGive 2 is modified
instead.

I can tweak drop amounts and probabilities as needed for balance...

This only affect sneakable actors right now.  I could apply this to
normal actors as well, so they would always drop their weapon with the
number of rounds that their internal reload counter was tracking, then
ammo and whatever else at random, but the enemies use half a dozen
different placeholder actors, so I'd need to standardize those, first
(probably convert them to a simple int counter instead of giving
inventory).