Alarm Tweaks/Fixes
- Added Z-height checking for player activation of alarm panels
- Changed delay for alarm panels so that a user-deactivated alarm panel
will wait for long enough that all enemies could return to idle before a
new activator is selected (so it's not just a continuous cycle of them
turning the alarm back on, even if they can't see you)
- Fixed alarms so that enemies within radius of the alarm will stay
active while the alarm is sounding instead of going idle afte rthe
normal countdown
- Tweaked dead body goal following to restore the ability to set a
custom state upon activation, so the Nazi's 'Alarm' state gets shown
again when he pulls the alarm.
C2M1: Removed extraneous monster blocking lines that were keeping the
enemies from activating the alarms in the guillotine room, and moved the
monster spawner off of the steps so the enemies' feet weren't stuck in
the floor.