Added ability for enemies to "see through" windows
- Really, this lets them see through any double-sided line that is
tagged as "block everything" - they'll see the player and go active, but
won't open fire.
- Implemented for both normal 'Nazi'-class actors and for sneakables.

Please test and let me know if there are any gameplay oddities as a
result of this!  It performs as expected in C2M1 as far as I can tell.

- Unrelated tweak: also changed alarm panels to only select Nazis that
can use walls as activators (the dogs could get selected before...).