Nitpicks
DECORATE & DEFS:
:new: Moved BloodSquirt on RavenBlood actor and removed useless space
for translated bloodtypes
:new: Fixed also decals for BloodPool2 actor though (DSPOT4-5 doesn't
exist so replaced with 1 and 2)

:notebook: Ceiling pools have been tested and they work perfectly (you
can try by killing an enemy with a rifle aiming to the ceiling, on a
64mp tall sector for example)
:warning: This is what's left after my attempt to fix GibSpray actor not
correctly being translated on XDeath states... Just spawn a mutant and
destroy it with kill Mutant CCMD, or kill it with a Nebelwerfer...
Noticed that it could be probably fixed on the Base actor zscript (the
XDeath state), but how? @AFADoomer Any ideas or suggestions?